rename core to common, mcp-server, to mcp, agent-server to agent
