+++ package.json
@@ -12,9 +12,9 @@
     "start:debug": "bun run --filter @browseros/server start:debug",
     "test": "bun test",
     "test:all": "bun test --workspace",
-    "test:core": "bun run --filter @browseros/core test",
+    "test:common": "bun run --filter @browseros/common test",
     "test:tools": "bun run --filter @browseros/tools test",
-    "test:mcp": "bun run --filter @browseros/mcp-server test",
+    "test:mcp": "bun run --filter @browseros/mcp test",
     "test:server": "bun run --filter @browseros/server test",
     "typecheck": "tsc --build",
     "build:binary": "mkdir -p dist && bun build --compile packages/server/src/index.ts --outfile dist/browseros-server --minify",
