+++ packages/common/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@browseros/core",
+  "name": "@browseros/common",
   "version": "0.0.1",
-  "description": "Core utilities for BrowserOS server",
+  "description": "Common utilities for BrowserOS server",
   "type": "module",
   "main": "./src/index.ts",
   "exports": {
