+++ packages/common/src/PageCollector.ts
