+++ packages/mcp/tsconfig.json
@@ -9,5 +9,5 @@
   },
   "include": ["src/**/*", "tests/**/*"],
   "exclude": ["node_modules", "dist/**/*"],
-  "references": [{"path": "../core"}, {"path": "../tools"}]
+  "references": [{"path": "../common"}, {"path": "../tools"}]
 }
