+++ packages/server/tsconfig.json
@@ -10,8 +10,8 @@
   "include": ["src/**/*", "tests/**/*"],
   "exclude": ["node_modules", "dist/**/*"],
   "references": [
-    {"path": "../core"},
+    {"path": "../common"},
     {"path": "../tools"},
-    {"path": "../mcp-server"}
+    {"path": "../mcp"}
   ]
 }
