+++ packages/tools/tsconfig.json
@@ -9,5 +9,5 @@
   },
   "include": ["src/**/*", "tests/**/*"],
   "exclude": ["node_modules", "dist/**/*"],
-  "references": [{"path": "../core"}]
+  "references": [{"path": "../common"}]
 }
